+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Change Log
 
 - [beta] Fix multi-cursor behavior of deleting to the beginning of visual line by `⌘⌫`.
 - [beta] Fix multi-cusor movement when encounter a surrogate pair.
+- [beta] Fix an issue where a selected range and an insertion point could overlap.